pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-detection-research"
version = "0.1.0"
description = "Adversarial robustness testing of AI text detectors"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "pandas>=2.0",
    "scipy>=1.11",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "jupyter>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
